+++ HISTORY.rst
@@ -9,6 +9,7 @@ develop
 - Show relative timestamps in files browser, file view and revisions view
 - Group folders on top in files browser
 - Sort folder and file names case insensitive
+- Use pixel size from results as default pixel size in results downloader
 
 
 1.1.21 (unreleased)